requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-phonenumber-field==8.0.0 
stripe==10.8.0
